# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ReniAgent — Telcel Sales Agent API
  description: >
    API del agente de ventas Telcel (ReniAgent). Atiende conversaciones de
    migración de plan por WhatsApp o por integración directa (`/chat`),
    administra campañas de contacto masivo, recibe el webhook de WhatsApp
    (Meta Cloud API) y el webhook de estatus de provisioning de Telcel.


    La mayoría de los endpoints requieren el header `x-api-key` (ver
    `ApiKeyHeader` en `securitySchemes`). Las excepciones están marcadas
    explícitamente en cada operación (`security: []`).
  version: "2.0.0"
servers:
  - url: https://o43h25qtjz656v3eohaj4mubua.apigateway.mx-monterrey-1.oci.customer-oci.com/telcel
    description: OCI API Gateway — ambiente productivo
  - url: http://localhost:8000
    description: Entorno local de desarrollo
...
paths:
...
  /campaign:
    post:
      tags: [Campañas]
      summary: Envía el template de campaña a una lista de líneas
      description: >
        Busca cada línea en la tabla `clientes`, arma la sesión inicial y
        envía el template de WhatsApp correspondiente al plan elegible del
        cliente (o texto plano si no hay plan elegible). Guarda la sesión
        bajo el `wa_id` canónico devuelto por Meta para que el webhook la
        encuentre al recibir la respuesta.
      operationId: sendCampaign
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignRequest'
            examples:
              phone_numbers:
                summary: Contrato real del API Gateway en producción
                value:
                  phone_numbers: ["2721140994"]
              lineas:
                summary: Contrato interno original (alias equivalente)
                value:
                  lineas: ["5555555555", "5555555556"]
      responses:
        "200":
          description: Resultado por línea — enviadas y fallidas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchResult'
        "401":
          $ref: '#/components/responses/InvalidApiKey'
        "422":
          $ref: '#/components/responses/ValidationError'
        "503":
          $ref: '#/components/responses/CampaignDbUnavailable'
  /campaign/dispatch:
    post:
      tags: [Campañas]
      summary: Batch-dispatch de campaña (usado por la UI Streamlit)
      description: >
        Para cada `linea`, busca el cliente en la BD, envía el template de
        WhatsApp y actualiza el registro `campana_clientes` correspondiente
        (`mark_enviado`/`mark_fallido`).
      operationId: campaignDispatch
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchRequest'
            example:
              campana_id: 1
              lineas: ["5555555555", "5555555556"]
      responses:
        "200":
          description: Resultado por línea.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchResultSimple'
        "401":
          $ref: '#/components/responses/InvalidApiKey'
        "422":
          $ref: '#/components/responses/ValidationError'
        "503":
          $ref: '#/components/responses/CampaignDbUnavailable'
  /campaign/dispatch/lote:
    post:
      tags: [Campañas]
      summary: Envía al siguiente lote de clientes pendientes de una campaña
      description: >
        Resuelve automáticamente hasta `limite` líneas con
        `estado_envio='pendiente'` para la campaña, ordenadas por fecha de
        alta, y reusa la misma lógica de `/campaign/dispatch`.
      operationId: campaignDispatchLote
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchLoteRequest'
            example:
              campana_id: 1
              limite: 100
      responses:
        "200":
          description: Resultado por línea, o `detail` si no había pendientes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchResultSimple'
              examples:
                sinPendientes:
                  summary: No hay clientes pendientes
                  value:
                    sent: []
                    failed: []
                    detail: "No hay clientes pendientes para esta campaña"
        "401":
          $ref: '#/components/responses/InvalidApiKey'
        "422":
          $ref: '#/components/responses/ValidationError'
        "503":
          $ref: '#/components/responses/CampaignDbUnavailable'
  /campaign/migrate-orden:
    post:
      tags: [Campañas]
      summary: Agrega/backfillea las columnas 'orden' y 'estado' en 'clientes'
      description: Migración idempotente — segura de correr varias veces. Pensada para bases de datos creadas antes de que existieran estas columnas.
      operationId: campaignMigrateOrden
      security:
        - ApiKeyHeader: []
      responses:
        "200":
          description: Resultado de la migración.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrateOrdenResult'
        "401":
          $ref: '#/components/responses/InvalidApiKey'
        "503":
          $ref: '#/components/responses/CampaignDbUnavailable'
  /campaign/migrate-plan-seleccionado:
    post:
      tags: [Campañas]
      summary: Agrega la columna 'plan_seleccionado' en 'clientes'
      description: Migración idempotente.
      operationId: campaignMigratePlanSeleccionado
      security:
        - ApiKeyHeader: []
      responses:
        "200":
          description: Resultado de la migración.
          content:
            application/json:
              schema:
                type: object
                properties:
                  columna_agregada:
                    type: boolean
                    example: false
        "401":
          $ref: '#/components/responses/InvalidApiKey'
        "503":
          $ref: '#/components/responses/CampaignDbUnavailable'
  /campaign/import-clientes:
    post:
      tags: [Campañas]
      summary: Importa/actualiza clientes desde un CSV
      description: >
        Formato de CSV igual al de `docs/Masivo_backup.csv`. Si se manda
        `campana_id`, vincula las líneas importadas a esa campaña
        (`estado_envio='pendiente'`). Si `reemplazar=true`, borra TODOS los
        clientes y vínculos existentes antes de importar.
      operationId: campaignImportClientes
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Archivo CSV con, mínimo, la columna `linea`.
                campana_id:
                  type: integer
                  nullable: true
                  description: Si se manda, vincula las líneas importadas a esta campaña.
                reemplazar:
                  type: boolean
                  default: false
                  description: Si es true, borra todos los clientes existentes antes de importar.
      responses:
        "200":
          description: Resultado de la importación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportClientesResult'
        "400":
          description: El CSV está vacío o no tiene la columna `linea`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              example:
                detail: "El CSV no tiene la columna 'linea'."
        "401":
          $ref: '#/components/responses/InvalidApiKey'
        "500":
          description: Error inesperado importando el CSV.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        "503":
          $ref: '#/components/responses/CampaignDbUnavailable'
  /campaign/clientes:
    get:
      tags: [Campañas]
      summary: Lista TODOS los clientes (sin filtrar por campaña)
      operationId: campaignGetAllClientes
      security:
        - ApiKeyHeader: []
      responses:
        "200":
          description: Lista completa de clientes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientesListResponse'
        "401":
          $ref: '#/components/responses/InvalidApiKey'
        "503":
          $ref: '#/components/responses/CampaignDbUnavailable'
    delete:
      tags: [Campañas]
      summary: Borra TODOS los clientes y sus vínculos de campaña
      description: Irreversible. Requiere `?confirm=true` para evitar disparos accidentales.
      operationId: campaignDeleteAllClientes
      security:
        - ApiKeyHeader: []
      parameters:
        - name: confirm
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Conteo de registros borrados.
          content:
            application/json:
              schema:
                type: object
                properties:
                  clientes_borrados:
                    type: integer
                  campana_clientes_borrados:
                    type: integer
                example:
                  clientes_borrados: 120
                  campana_clientes_borrados: 45
        "400":
          description: Falta `?confirm=true`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              example:
                detail: "Falta ?confirm=true — esta acción borra TODOS los clientes."
        "401":
          $ref: '#/components/responses/InvalidApiKey'
        "503":
          $ref: '#/components/responses/CampaignDbUnavailable'
  /campaign/{campana_id}/clientes:
    get:
      tags: [Campañas]
      summary: Lista los clientes vinculados a una campaña
      description: Join de `clientes` + `campana_clientes`, útil para verificar una importación.
      operationId: campaignGetClientes
      security:
        - ApiKeyHeader: []
      parameters:
        - name: campana_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Clientes vinculados a la campaña.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampanaClientesListResponse'
        "401":
          $ref: '#/components/responses/InvalidApiKey'
        "422":
          $ref: '#/components/responses/ValidationError'
        "503":
          $ref: '#/components/responses/CampaignDbUnavailable'
  /reportes/campana-clientes:
    get:
      tags: [Campañas]
      summary: Reporte de trazabilidad de campaña(s)
      description: >
        Devuelve el contenido de `campana_clientes` con los datos del cliente
        asociado. Si `campana_id` se omite, trae las filas de todas las
        campañas.
      operationId: reportesCampanaClientes
      security:
        - ApiKeyHeader: []
      parameters:
        - name: campana_id
          in: query
          required: false
          description: Si se omite, trae todas las campañas.
          schema:
            type: integer
        - name: full
          in: query
          required: false
          description: Incluir `historial_conversacion` y `session_data_snapshot`.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Filas de trazabilidad.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CampanaClienteReporte'
        "401":
          $ref: '#/components/responses/InvalidApiKey'
        "503":
          $ref: '#/components/responses/CampaignDbUnavailable'

  # ── Webhook WhatsApp ────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
      description: API key compartida (variable de entorno `API_KEY`). Si `API_KEY` no está configurada del lado del servidor, no se exige.
    ApiKeyQuery:
      type: apiKey
      in: query
      name: api_key
      description: Misma API key que `ApiKeyHeader`, pero enviada como query param — solo la usa `/logs`.
  responses:
    InvalidApiKey:
      description: API key ausente o inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            detail: "Invalid API key"
    ValidationError:
      description: Cuerpo de la petición inválido (validación automática de FastAPI/Pydantic).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    CampaignDbUnavailable:
      description: >
        La base de datos de `campaign_app` no está disponible en este
        despliegue (import opcional fallido) — el agente conversacional
        sigue funcionando igual, pero los endpoints de campañas responden 503.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            detail: "Campaign DB not available: <traceback>"
  schemas:
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
          example: "Invalid API key"
...
    CampaignRequest:
      type: object
      description: >
        Acepta 'lineas' o 'phone_numbers' (alias equivalente, es el nombre
        real que usa el contrato del API Gateway en producción). Al menos
        uno de los dos debe traer valores.
      properties:
        lineas:
          type: array
          items:
            type: string
        phone_numbers:
          type: array
          items:
            type: string
    DispatchRequest:
      type: object
      required: [campana_id, lineas]
      properties:
        campana_id:
          type: integer
        lineas:
          type: array
          items:
            type: string
    DispatchLoteRequest:
      type: object
      required: [campana_id]
      properties:
        campana_id:
          type: integer
        limite:
          type: integer
          default: 100
    DispatchResult:
      type: object
      description: Resultado de `/campaign` — un elemento por línea enviada o fallida.
      properties:
        sent:
          type: array
          items:
            type: object
            properties:
              phone_number:
                type: string
              wa_id:
                type: string
                description: Número canónico devuelto por Meta — clave usada por el webhook.
              sent_template:
                type: boolean
              template_name:
                type: string
                nullable: true
        failed:
          type: array
          items:
            type: object
            properties:
              phone_number:
                type: string
              reason:
                type: string
...
security:
  - ApiKeyHeader: []